Updated openstack/openstack Project: openstack-infra/system-config  bc4cdb2d2f908d8b963a7a47223b20e72af68fb6
Added AFS keytab to wheel slave

This adds a hiera-loaded keytab for the wheel slaves. It may
be used when writing wheels to the AFS share directory.

Note: This will require the creation of a keytab for the
wheel mirror, as well as appropriate folders and access rights
on AFS.

Change-Id: I81c39d420d8ac70def57949ea0d4c323b8797086