Updated openstack/openstack Project: openstack/neutron  144891577debaf7c33b348c307aa7ad45a4d3842
DVR: avoid race on dvr serviceable port deletion

Please see race details in the bug report.
The fix is to check for dvr routers that needs to be removed
after actual port deletion.
As a bonus we can get rid of dvr logic in ml2 plugin delete_port().

It's hard to come up with a test here as race can only be
reproduced with multiple API workers.

Closes-Bug: #1538163
Change-Id: I7ab1b0c18e5daa6f35a37370b9be79e689e2ba50