Updated openstack/openstack Project: openstack/tempest  f8235411bf7c8f1a6ea80843d8c3634c676a4da5
Consolidate the ssh code

The entire mechanism for ssh validation is a mess. The fact that there
is a parameter which is server_or_ip, so you never know if it's an
object or an address, and that every one of 7 layers need to handle
either case, makes doing any enhancements here for failure detection
impossible.

This change implements a get_server_ip method that gets the fixed or
floating IP address before passing it to the remote client.

Change-Id: I76d818bf9e1bd26b1e499464fc9213764afee279