Updated openstack/openstack Project: openstack/nova  6fd6dd7c13e71c85c6f0f2d0591b77b81151e8fe
remove the wrong param of fake_db_migration initiation.

In test_get_migrations, fake_db_migration is initated with a
wrong param.

There is no uuid field in migration db.
It should be instance_uuid.

The test case does not verify this param.
If we remain it here, the developers may make wrong impression
on migration db struct.

This patch removes it.
So do not mislead the developers.

Change-Id: I35ce83a746391cdc19d26769841b0ae123e2c2be