Updated openstack/openstack Project: openstack-infra/project-config  204faf821528141df227616e9eb04a13d7b4122e
Adding gerritbot and accessbot support for openstack-product

The openstack-product IRC channel will be used by the OpenStack Product
Working Group for additional conversations when the meetings take longer
than our allocated time on the openstack meeting channel.  The product
working group will also be in the channel so that we can answer any
questions related to our working group and/or user stories.

Change-Id: Id3779c5dbd750d3d34a2866d514170de12dea073