Updated openstack/openstack Project: openstack/tempest-lib  ce625da05c364d80cf1b69c7e48867cc9d4a44eb
Added __init__.py to tests/services/identity/

This is necessary for the tests to actually be run.

Also use sort_keys=True for json.dumps() in both the auth() methods and the
unit tests that compare the expected and actual request bodies. Otherwise
the order of the keys is unpredictable and the tests fail randomly.

Change-Id: Ieb27be22367b419326939c4261388286e6c996c4