Updated openstack/openstack Project: openstack/oslo.config  ad461d1967f4d7e2d0dcfb27ccd99eca0e792b90
Removal of bandit.yaml

Newer versions of Bandit do not require a bandit.yaml. The copy
of bandit.yaml included here was identical to the default ignoring
some style differences. The future direction of Bandit is to have
no need for a separate yaml. Everything will be command line driven.

Change-Id: I9a31e478605ba07fd7602003566cbef2a12433f7