Updated openstack/openstack Project: openstack/ironic  7e734252e997d7eb7f76069857982c2fadaa3bc2
Extend root device hints to support device name

This patch is extending the root device hints to support the device name
as input.

It's important to note that for SATA, SCSI and IDE disk controllers
the name is not recommended to be used because the order in which the
device nodes are added in Linux is arbitrary, resulting in devices like
/dev/sda and /dev/sdb switching around at boot time. The documentation
was updated to inform operators about this problem.

Depends-On: I48d6456c75bbe6ddf16ac6561e5461ca51eb9c37
Change-Id: I564d68ab1c66505195745c34e16e829b45a00a75
Closes-Bug: #1526732