Updated openstack/openstack Project: openstack/oslo.messaging  1cc3b9c80e948b5f7052ac76f03d31a143269cf6
Switches pika driver to eager connection to RabbitMQ

In previous implementation PikaPoller returned by PikaDriver.listen
is not connected to RabbitMQ. Connection is established only when
executor calls first poll method

This behaviour is different in comparison with kombu driver and causes
neutron rpc workers hanging.

I tried to fix it there:
https://review.openstack.org/#/c/270832/

But oslo.service team don't accept this workaround
The main problem is that neutron uses olso libraries in wrong way and
it is necessary to change it in neutron. This patch only mask this
problem

Change-Id: Idd6b692bcf2025fc2a3b37b27f4d1c1b01922fdf