Updated openstack/openstack Project: openstack/heat  260eacd68b073c190c8307802fe4b5317542b652
Remove implementation_signature check

Remove the unnecessary checking of implementation signature of
current resource with the prev_resource in needs_update.
It seems prev_resource is always resource from the backup_stack
and there would be no difference.

If there is a change of implementation class, it would have been
already handled[1]

[1]
https://github.com/openstack/heat/blob/master/heat/engine/update.py#L139-L140

Change-Id: I1df84986cbc4bae7bcdbaae8b5aacbfa7adaec32