Updated openstack/openstack Project: openstack/barbican  2f05bb1faea32e8ddea37935cee24026455d227b
Use host href for version discovery

In barbican we already use the host_href coming from the configuration
to return the paths the user needs to use to retrieve their respective
secrets. This, however, is not the case for version discovery. The
version discovery (json-home) implementation done in barbican will
return the path by which barbican was accessed. This is problematic
because it might not be the path that we want to expose barbican
through. On the other hand, we might want to enforce the usage of TLS
for accessing barbican; This will not be the case if there is a
possibility of accessing barbican through plain HTTP (in the case where
the loadbalancer is not explicitly blocking so).

As a solution, the json-home output will be the value that was
configured in the host_href option. This will then return endpoints for
keystone discovery that coincide with the paths that are returned with
entities.

Change-Id: I69f556c09f072b47e9f8e776ccd4db209b25485f