Updated openstack/openstack Project: openstack/python-openstackclient  3763d75dbad477ad654daec321c0291550152873
Network: Abstract get_body() out to be a private helper.

get_body() is needed in each network files to construct a dict
to pass to sdk proxy. And it is also used by several functions
in each file. So define it as a file level private helper function.

The unified prototype should be:

    def _get_attrs(client_manager, parsed_args):

1. The name, in sdk, the parameter passed to proxy is named "attrs".
   And it is a private method. So let's call it _get_attrs().
2. The parameters, besides parsed_args, when we deal with project
   and project_domain, we have to make use of identity_client. So
   let's pass in the client manager.

Change-Id: Ib044ebd4ddedbcd805f46334a7fe99e4ebb5b249