Updated openstack/openstack Project: openstack/taskflow  61efc31e965b8defac84aa494095fd3734f739eb
Handle cases where exc_args can't be serialized as JSON in the WBE

First try to get them, but if they aren't able to be serialized,
just return the failure without them.  This only affects the WBE
that serializes failures as part of the JSON response that is sent
over the wire from the worker process to the WBE engine.

Change-Id: I86e3255b612bc15097aabe63a684cf8b8808e61b