Updated openstack/openstack Project: openstack/cinder  9299e7aebea42095fe1fa09b89d7297b5a53826e
EMC VMAX - Fix for last volume in VMAX3 storage group

This patch fixes bug in the deletion the last volume in a
storage group for the VMAX3. If a masking view did not exist
we tried to get an instance of it which threw an exception.
Now we handle both scenarios correctly.

Closes-Bug: #1512796

Change-Id: Ideb2aaffe4fbc1ae2137d569d7b137ac0edef573