Updated openstack/openstack Project: openstack-infra/project-config  8ab09821a0109f09e240555c17cd793376b34484
add a job to send automated announcements of releases

Define a new job {name}-announce-release to send announcements of new
releases and associate it with the "release" queue to be run after the
release is completed and uploaded to PyPI (for libraries) or the tarball
server (for server projects). We don't want pre-releases (alphas, betas,
and release candidates) to have announcements generated automatically.

The job uses a script in the openstack-infra/release-tools repository to
actually generate and send the email. This is part of the same set of
tools we've been using for manual release announcements.

Spec: release-automation
Story: 2000421
Change-Id: I8106f8449d6c41da6b799aee484e81ffebb498bf
Signed-off-by: Doug Hellmann <doug@doughellmann.com>
Signed-off-by: Jeremy Stanley <fungi@yuggoth.org>