Updated openstack/openstack Project: openstack/diskimage-builder  1dcd65fb9e57b84abda664484418df340af07147
Print unparsable file in pkg-map

As you can see in the CI failures for
Ib11b9df84b593ab25232729a570c812f1b4b8774, you can not see what file
is causing the problems when the parser fails.  Add a print, but raise
the error as it contains the cause.

Change-Id: I127ff7d57d2d898969195464c6e774d496e872e6