Updated openstack/openstack Project: openstack-infra/project-config  8daaa758a2f3ec4f09307e66610ce34394242160
Remove ML reporting for periodic-qa jobs

This commit removes the email reporting from the periodic-qa jobs.
With openstack-health being a thing now we no longer need to report
to the deprecated ML. This is the first step in removing the
openstack-qa list, which has been a goal for a long time.

Change-Id: I9e5f6e2aabc1a223391a9898ec578d6abd5d13c7