Updated openstack/openstack Project: openstack/governance  ecd00ff4d3ca6f345e5dfce0242d85f8386f4f94
neutron: Change release model for Octavia, OVN, and NSX

Update the release model for networking-ovn, octavia, and vmware-nsx
to be "release:independent." These projects have not released anything
during Mitaka yet, so the release tag of "cycle-with-milestones" is no
longer accurate. We can change this tag back to "cycle-with-milestones"
during Newton if the teams desire.

Change-Id: Ib50d6ba766075c64ccd13824096278b165223a05
Signed-off-by: Kyle Mestery <mestery@mestery.com>