Updated openstack/openstack Project: openstack/heat  7c68e664e90527378aca5ee0ab298d800f40beda
Translate all properties during update

This adds a translate call for before properties during update, so that
we can retrive the proper keys when checking them in new resources. This
also handles the network key in the Port resource to not update if
network happens to be in the changed properties.

Change-Id: I4d76b5f60f454babbe60832d9e81912c13d53b9c
Closes-Bug: #1541516