Updated openstack/openstack Project: openstack-infra/storyboard  cbe35d606a067c757ba7200e5e1181cd9528e331
Don't add non-existent items when resolving lists

When a task/story is deleted, the WorklistItem for that task/story
is not deleted. Currently, this raises an exception when resolving
the items of the worklists containing that task/story.

This patch fixes that by skipping WorklistItems whose items no longer
exist. In the long run, we should just not delete tasks/stories and
archive them instead.

Change-Id: I238bc4b86b5358fd11080be6ebfd9b76b09753f4