Updated openstack/openstack Project: openstack/glance  8781f64db2a7ce1edad4018517aabc50a5880b36
Prevent user to remove last location of the image

If the last location of the image is removed, image transitions back to queued.
This allows user to upload new data into the existing image record. By
preventing removal of the last location we prevent the image transition back to
queued.

This change also prevents doing the same operation via replacing the locations
with empty list.

SecurityImpact
DocImpact
APIImpact

Change-Id: Ieb03aaba887492819f9c58aa67f7acfcea81720e
Closes-Bug: #1525915