Updated openstack/openstack Project: openstack/neutron  4e1b14671d8f156b37b9c6eb24e1670c913f9816
Only prevent l3 port deletion if router exists

This patch modifies the prevent_l3_port_deletion method to
actually look up the router_id in the device_owner field to
confirm that the router exists before claiming the port is
in use. This will allow users to delete ports that may have
been orphaned due to race conditions in the cleanup of router
interfaces.

Partial-Bug: #1540271
Change-Id: Ieffe632f3f3098baf202d3795ab5182982e234bd