Updated openstack/openstack Project: openstack/neutron-lbaas  e446f976889d587ff0f8455c714fc334e3dabe2f
Make sample web server close connections

The sample web server, using nc, currently does not close the
connection after serving the content.  This can cause clients to
hang waiting for the connection to close after the content has been
served.  This patch includes the content length in the response
so that the connection will close.

Change-Id: I9b7d120099c2652a2a91e3471e3152c1b02ce7f6