Updated openstack/openstack Project: openstack/nova  355fc69029e7e86a4f81e8464a3054e171961f44
Adding volume operations for shelved instances

At the moment Nova doesn't allow to perform attach detach operations on
instances in shelved or shelved_offloaded state, those operations can be
enabled by allowing them at the compute API for those states
This change is to put in place all the code change necessary to manage
the attach and detach operations for shelved/shelved_offloaded
instances.
When an instance is in shelved_offloaded state it is not on a compute
host so we don't need to physical attach/detach the volume from the
instance but we need to take care of accounting this information in the
Nova and Cinder DB.
We are not enabling this new feature in this change but the
required API changes will be in a following one.
For that reason this commit message doesn't require any specific TAG.

blueprint volume-ops-when-shelved

Change-Id: I27d983a092910a6eb54e2833a8ff4d568ac00cdf