Updated openstack/openstack Project: openstack/manila  7965b7bd39832bb8fdaa4329faa9a79aeed3d240
Fix Mutable default argument

Python’s default arguments are evaluated once
when the function is defined, not each time the function is called.

This means that if you use a mutable default argument
and mutate it, you will and have mutated that object
for all future calls to the function as well.

more details about this wrong usage here:
http://docs.python-guide.org/en/latest/writing/gotchas/#mutable-default-arguments

Closes-Bug: #1534974
Change-Id: I52140f8dc268541d44f54721da039c631b28d98a