Updated openstack/openstack Project: openstack/keystone  6fe74951ac1c41931e370193387722fcc4e49dcb
Add tests for role management with v3policy file

In the upcoming patches for domain specific roles, we will modify
the policy rules for role manipulation (e.g. a domain admin in
domainA should not be able to manipulate domain specific roles in
domain B).  However, it turns out we don't test the currect role
protection with the v3policy sample at all!

This patch therefore adds this testing, so that in the follow-on
patches for domain specific roles we can show that we have not
affected the rights to manipulate global roles.

Partially Implements: blueprint domain-specific-roles

Change-Id: Ifdce381857fd7ad4cc2e5183d136907f8f3bf561