Updated openstack/openstack Project: openstack-infra/system-config  3b1e79f732bd1bd1623090fbe3d8b4736d34b06a
Increase the HTTP incoming connection queue

Upstream Gerrit confirms that the default value for the httpd.maxQueued
value was set too low.  The fix[1] for it was to change the default to
200 however the fix is only in version 2.12+ not in 2.11.

clarkb confirmed that we are seeing the same error (stack trace linked
in 1st ml thread in upstream fix[1]) in review.o.o logs.  This likey
indicates that VM starvation (due to GC) is making Gerrit to slow to
process incoming client requests so it pushes more requests into the
queue than it can handle.  Once the queue is exhausted the requests
get rejected with 500 errors.  Increasing the max number of client
connections may help alleviate this situation until Gerrit can catch
up.

[1] https://gerrit-review.googlesource.com/#/c/70627

depends-on: Iea202fcfe7af4a9d5b4c3a360124c24649b1133b
Change-Id: I8d6c8d9ec6348fc6b950744237e527133a625bbe