Updated openstack/openstack Project: openstack/ironic  7e1690768c4c2bcb18594f542937ae8c2c214fab
Fix genconfig "tempdir" inconsistency

This patch is mocking the tempfile.gettempdir() method as part of the
sample config generation. Different hosts may return a different
directories on that call (some making very little sense), so it's
important that we make it consistent accross hosts to avoid conflicts
when generating the sample configuration file.

Closes-Bug: #1542255
Change-Id: Id5d5db766f3017bcb7c571805d94a48acf71de9b