Updated openstack/openstack Project: openstack/neutron-fwaas  39012a787edcdea17e28abe7dd7a3bc647edd4ce
Add an explicit BRANCH_NAME to tox_install.sh

tools/tox_install.sh exists to deal with installing neutron without
neutron existing as a pypi package.

The way this operates causes problems in the periodic jobs on the stable
branches.  Make it explicit which branch to install in the gate
(zuul-cloner) and via pip.

This is done with a BRANCH_NAME variable to make it trivial to change
once "master" is tagged as stable.

Change-Id: I6a5b63fa82a6a7480adadd2f65463619fd077f7f