Updated openstack/openstack Project: openstack/neutron  e74ccaec9c0c8ffe5f9342e4b7ec5f7d7ac5db2b
Add separate transport for notifications

In oslo.messaging 2.9.0, the RPC and notification buses were decoupled
into separate transport. This changes over the Neutron notifier to use
the notification transport provided from oslo.messaging instead of the
original transport (which is now used for RPC). By default, the
notification transport will pull in the original transport if nothing is
configured in oslo_messaging_notification, so this can be an in-place
replacement without impacting upgrades or configurations.

This functionality was added to oslo.messaging to address bug #1504622

Change-Id: I89c8a84c81079677088cc7d656a0bcb5e323de4d