Updated openstack/openstack Project: openstack/neutron-vpnaas  e6f643f2a1099dfeab37c5865826797b5c6e78e5
Add an explicit BRANCH_NAME to tox_install.sh

tools/tox_install.sh exists to deal with installing neutron without
neutron existing as a pypi package.

The way this operates causes problems in the periodic jobs on the stable
branches.  Make it explicit which branch to install in the gate
(zuul-cloner) and via pip.

This is done with a BRANCH_NAME variable to make it trivial to change
once "master" is tagged as stable.

Change-Id: I484637d2502f276313edc2279078a200ebc2689a