Updated openstack/openstack Project: openstack/nova  9f7eebec2bafd6ba98475567ebbd6139b1a43416
Add a TODO to make ComputeNode.cpu_info non-nullable

The cpu_info field in ComputeNode is nullable but the field in the
schema is non-nullable, so attempts to create a ComputeNode object
without specifying cpu_info fail at the db layer with:

  IntegrityError: NOT NULL constraint failed: compute_nodes.cpu_info

Unfortunately, we can't change the field to non-nullable until
version 2.0 of ComputeNode because it could be possible for old
nodes to send null cpu_info over the wire to new nodes which would
fail during deserialization. This adds a TODO to make the field
match the schema at version 2.0.

Change-Id: I013f05b6eabd5e7cafa962d61ad5356f8bde45fc