Updated openstack/openstack Project: openstack/nova  559b8aaea847fa18f1fbe314ed7b81305f6096e0
Optimize servers path by pre-joining numa_topology

Paths like server delete pass the instance down to the compute node,
which expects numa_topology to be joined. This currently causes a
lazy-load on every such operation. In a recent tempest run, numa_topology
was lazy-loaded 134 times (not all of which are in server delete,
necessarily).

This patch adds it to the list of attributes we pre-query when
starting an operation like delete, update, etc.

Change-Id: Ia4317fb1690535f467e1b295fbab64c9ec2f6f76
Partial-Bug: #1540526