Updated openstack/openstack Project: openstack/horizon  0ab863e9652e0799ba23b50df610e03ebb7a8e49
Removing container ID element

Removing the element, as it was not really needed at all.
Also, removed a debug statement that was discovered in
material.hamburger.js while refactoring for this change.

Added a 'container-fluid' class to the content body
as well, because it is not possible to configure Horizon
as a non-fluid layout without a container element around
the first 'row' element.

Change-Id: I097ea31f991e8d15987b7cf54b77a309d6155771
Closes-bug: #1368924