Updated openstack/openstack Project: openstack/glance  c82f5798014d2f51663bf4c524da64956d7e1132
Make cache config options clearer

Currently the cache config options are unclear.

The ``image_cache_stall_time`` is made to sound like it applies to all
images rather than just incomplete images. This could give users the
wrong impression of this option.

The other commands are updated to make them easier to read and follow
the capitalisation of the usages.

DocImpact
Change-Id: Ia7a81c37e626eddb50c59a5ed854abff52bc89fe